### 2024-01-17 #622 @Rekyt
funbiogeo: a data processing and visualization package for functional biogeography
Submitting Author Name: Matthias Grenié
Submitting Author Github Handle: @Rekyt
Repository:  https://github.com/FRBCesab/funbiogeo
Submission type: Pre-submission

---

```
Helps users with analyses in functional biogeogeography by loading
and combining data, computing trait coverage, as well as computing
functional diversity indices, drawing maps, correlating them with the
environment, and upscaling assemblages.
```

The package aims to provide tools to easily visualize, filter, and analyze data in functional ecology and functional biogeography: species-traits matrices, site-species matrices. It can be used both as a pre- and post-processing tool in functional ecology.

There are three main targets for the package:
- Newcomers to the field of functional ecology (= trait ecology), most packages focus on computing functional diversity indices, but there exists no tool to help provide a standard workflow to guide your analyses. These can be students, researchers, etc.
- Experts in functional ecology who are looking for quick and easy visualizations and data exploration tools.
- Applied scientists who would like to use the functional approach but are unsure how to proceed and would like some advanced feature like data upscaling.

To our knowledge, all other packages focus on computing functional diversity indices and creating functional spaces, but none are focused on the steps prior and after these.

N/A

 N/A
