### 2024-01-17 b225109
doc: use bootstrap 5 in pkgdown
