### 2024-01-17 efa579a
doc: remove unnecessary badges
