### 2024-01-17 f57bac2
ci: remove action to knit md files
