### 2024-01-17 a4ef90f
Upgrade bootstrap version (#100)
* doc: use bootstrap 5 in pkgdown

* doc: remove unnecessary badges

* ci: remove action to knit md files

---------

Co-authored-by: Nicolas Casajus <ahasverus@users.noreply.github.com>
