+++ pkgdown.yml
@@ -6,5 +6,8 @@ articles:
   funbiogeo: funbiogeo.html
   long-format: long-format.html
   upscaling: upscaling.html
-last_built: 2024-01-16T21:43Z
+last_built: 2024-01-17T13:59Z
+urls:
+  reference: https://frbcesab.github.io/funbiogeo/reference
+  article: https://frbcesab.github.io/funbiogeo/articles
 
