### 2024-01-17 f89d522
doc: add contribution guidelines
