### 2024-01-17 47de748
doc: add contribution guidelines (#101)
