+++ README.Rmd
@@ -26,18 +26,31 @@ pkgload::load_all()
 [![License: GPL v2](https://img.shields.io/badge/License-GPL%20v2-blue.svg)](https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html)
 <!-- badges: end -->
 
+
+
+<p align="left">
+  • <a href="#overview">Overview</a><br>
+  • <a href="#features">Features</a><br>
+  • <a href="#installation">Installation</a><br>
+  • <a href="#first-steps">First steps</a><br>
+  • <a href="#long-form-documentations">Long-form documentations</a><br>
+  • <a href="#citation">Citation</a><br>
+  • <a href="#contributing">Contributing</a><br>
+  • <a href="#acknowledgments">Acknowledgments</a><br>
+  • <a href="#references">References</a>
+</p>
+
+
+
 ## Overview
 
-The package `funbiogeo` aims to help users with analyses in functional
-biogeography (Violle et al. 2014) by loading and combining data, computing
-trait coverage, as well as computing functional diversity indices, drawing maps,
-correlating them with the environment, and upscaling assemblages.
+The package `funbiogeo` aims to help users with analyses in functional biogeography (Violle _et al._ 2014) by loading and combining data, computing trait coverage, as well as computing functional diversity indices, drawing maps, correlating them with the environment, and upscaling assemblages.
 
-It is aimed at first-timers of functional biogeography as well as more
-experienced users who want to obtain quick and easy exploratory plots.
+It is aimed at first-timers of functional biogeography as well as more experienced users who want to obtain quick and easy exploratory plots.
 
 
-## Main Features
+
+## Features
 
 `funbiogeo` offers:
 
@@ -52,26 +65,6 @@ experienced users who want to obtain quick and easy exploratory plots.
 * Functions to easily "upscale" your data to coarser spatial resolutions.
 
 
-## Citation
-
-For the moment `funbiogeo` doesn't offer a companion paper. But if you happen
-to use it in your paper you can cite the package through:
-
-> Casajus N, Grenié M (2022). _funbiogeo: Functional Biogeography Analyses_.
-  R package version 0.0.0.9000, <https://github.com/frbcesab/funbiogeo>.
-
-or the BibTeX version:
-
-```
-@Manual{,
-    title = {funbiogeo: Functional Biogeography Analyses},
-    author = {Nicolas Casajus and Matthias Grenié},
-    note = {R package version 0.0.0.9000},
-    url = {https://github.com/frbcesab/funbiogeo},
-  }
-
-```
-
 
 ## Installation
 
@@ -84,21 +77,18 @@ remotes::install_github("FRBCesab/funbiogeo")
 ```
 
 
+
 ## First steps
 
-This section will show you some useful functions from `funbiogeo`.
-For a longer introduction please refer to the
-["Get started" vignette](https://frbcesab.github.io/funbiogeo/).
+This section will show you some useful functions from `funbiogeo`. For a longer introduction please refer to the ["Get started" vignette](https://frbcesab.github.io/funbiogeo/).
 
-The package contains default example data named `species_traits`,
-`site_species`, and `site_locations`. You can for example visualize to completeness of your trait dataset using the `fb_plot_species_traits_completeness()` function:
+The package contains default example data named `species_traits`, `site_species`, and `site_locations`. You can for example visualize to completeness of your trait dataset using the `fb_plot_species_traits_completeness()` function:
 
 ```{r plot-sp-tr-complete}
 fb_plot_species_traits_completeness(species_traits)
 ```
 
-One other useful visualization is to see the trait coverage of each trait across
[…]
