+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   funbiogeo: funbiogeo.html
   long-format: long-format.html
   upscaling: upscaling.html
-last_built: 2024-01-17T13:59Z
+last_built: 2024-01-17T14:24Z
 urls:
   reference: https://frbcesab.github.io/funbiogeo/reference
   article: https://frbcesab.github.io/funbiogeo/articles
