+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   funbiogeo: funbiogeo.html
   long-format: long-format.html
   upscaling: upscaling.html
-last_built: 2024-01-17T14:24Z
+last_built: 2024-01-18T15:02Z
 urls:
   reference: https://frbcesab.github.io/funbiogeo/reference
   article: https://frbcesab.github.io/funbiogeo/articles
