### 2024-01-18 417c058
Update CITATION.cff
