+++ CITATION.cff
@@ -22,7 +22,7 @@ authors:
   email: matthias.grenie@idiv.de
   orcid: https://orcid.org/0000-0002-4659-7522
 repository-code: https://github.com/frbcesab/funbiogeo
-url: https://github.com/frbcesab/funbiogeo
+url: https://frbcesab.github.io/funbiogeo/
 contact:
 - family-names: Casajus
   given-names: Nicolas
