### 2024-01-23 #622 @jhollist
@Rekyt the other editors and I are discussing your pre-submission and whether it is within rOpenSci scope.  Hope to have an answer for you soon.
