### 2024-01-23 #622 @Rekyt
> @Rekyt the other editors and I are discussing your pre-submission and whether it is within rOpenSci scope. Hope to have an answer for you soon.

Alright! Thanks for the update!
That was precisely the goal of the pre-submission inquiry 😉
