### 2024-01-25 #622 @jhollist
@Rekyt Sorry for the delay in getting back to you!

We have had a chance to take a look and discuss this.  I do think funbiogeo is out of scope for geospatial data and mostly out of scope for data munging.  However, we do think it is a pretty good fit for the [Exploratory Data Analysis (EDA) and Summary Statistics](https://stats-devguide.ropensci.org/overview.html#overview-eda) category for statistical software packages.

Given this, I would encourage a statistical software submission in the EDA category.

Most of our checks look good.  Only thing you'll need to add if you do move ahead with a full submission is the codemeta.json file.  Details on this are on https://devguide.ropensci.org/building.html?q=codemeta#creating-metadata-for-your-package.

Thanks!
