### 2024-01-25 #622 @Rekyt
@jhollist thanks for your help.

Thanks for the clarification of categories! I didn't realize it could be submitted as a statistical package.
In order to proceed to a full submission, should I edit this issue or close this one and open a new one?
